iotaProject/requirements.txt
boto==2.42.0
Django==1.10
django-storages==1.5.0
django-widget-tweaks==1.4.1
Pillow==3.3.0
psycopg2==2.6.2
Wand==0.4.3
